+++ scripts/migrations/migration_1.45_users.js
@@ -1,17 +1,21 @@
 const fs = require("fs");
+const path = require("path");
 const yargs = require("yargs");
-const crypto = require("crypto");
+const ULID = require("ulid");
+const { createHash } = require("crypto");
 
-const argv = yargs.alias("f", "file").describe("f", "Path to users file").demandOption(["f"]).alias("w", "write").describe("w", "Write to the file").boolean("w").help().argv;
+const argv = yargs.alias("c", "config").describe("c", "Path to config directory").demandOption(["c"]).alias("w", "write").describe("w", "Write to the file").boolean("w").help().argv;
 
-const usersFilePath = argv.file;
+const usersFilePath = path.resolve(argv.config + "/users/users.json");
 
 fs.readFile(usersFilePath, (_err, rawData) => {
     const data = JSON.parse(rawData);
     const lData = Object.entries(data);
-    const lNewData = lData.map(([_key, user]) => {
+    const lNewData = lData.map(([key, user]) => {
         if (!user.token) {
-            user.token = crypto.randomBytes(20).toString("hex");
+            const hashedLogin = createHash("sha256").update(key).digest("hex");
+            const ulid = ULID.ulid();
+            user.token = `sls-${ulid.toLowerCase()}${hashedLogin.substring(0, 5)}`;
         }
         return [user.login, user];
     });