+++ scripts/run-migrations.sh
@@ -1,0 +1,6 @@
+#!/bin/env sh
+
+for FILE in $(find scripts/migrations/ -type f -name "migration_*.js" | sort); do
+    echo "Run the migration $(basename $FILE)"
+    node $FILE -c config -w && echo "🎉 DONE"
+done